EventPage 컴포넌트화
